Made zoom parameters ini file settings
git-svn-id: https://svn.code.sf.net/p/deinterlace/code/trunk/DScaler@647 6e12ab15-cf2f-4ffe-bf36-45892a6a5b89

+++ DScaler/AspectRatio.h
@@ -177,10 +177,10 @@ typedef struct _aspectsettings
 	BOOL DetectAspectNow;
 
 	// Zoom
-	double xZoomFactor;
-	double yZoomFactor;
-	double xZoomCenter;
-	double yZoomCenter;
+	long xZoomFactor;
+	long yZoomFactor;
+	long xZoomCenter;
+	long yZoomCenter;
 } AspectSettingsStruct;
 extern AspectSettingsStruct aspectSettings;
